foc-ads-manager/requirements.txt
# Google Ads Agent — The Foundation of Change
google-ads>=25.1.0
python-dotenv>=1.1.0
schedule>=1.2.0
requests>=2.32.0
tabulate>=0.9.0
